+++ Gruntfile.js
@@ -20,7 +20,7 @@ module.exports = function (grunt) {
       , options: {
           specs: "spec/**/*.js"
         , vendor: "vendor/**/*.js"
-        , version: '2.0.0-rc3'
+        , version: '2.0.0'
       }
     }
   })