+++ spec/suites/jasmine-jquery-spec.js
@@ -11,7 +11,7 @@ describe("jasmine.Fixtures", function () {
 
   beforeEach(function () {
     jasmine.getFixtures().clearCache()
-    spyOn(jasmine.Fixtures.prototype, 'loadFixtureIntoCache_').andCallFake(function (relativeUrl){
+    spyOn(jasmine.Fixtures.prototype, 'loadFixtureIntoCache_').and.callFake(function (relativeUrl){
       this.fixturesCache_[relativeUrl] = ajaxData
     })
   })
@@ -32,18 +32,18 @@ describe("jasmine.Fixtures", function () {
         jasmine.getFixtures().read(fixtureUrl)
         jasmine.getFixtures().clearCache()
         jasmine.getFixtures().read(fixtureUrl)
-        expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(2)
+        expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
       })
     })
 
     it("first-time read should go through AJAX", function () {
       jasmine.getFixtures().read(fixtureUrl)
-      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(1)
+      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
     })
 
     it("subsequent read from the same URL should go from cache", function () {
       jasmine.getFixtures().read(fixtureUrl, fixtureUrl)
-      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(1)
+      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
     })
   })
 
@@ -193,7 +193,7 @@ describe("jasmine.Fixtures", function () {
       it("should go from cache", function () {
         jasmine.getFixtures().preload(fixtureUrl, anotherFixtureUrl)
         jasmine.getFixtures().read(fixtureUrl, anotherFixtureUrl)
-        expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(2)
+        expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
       })
 
       it("should return correct HTMLs", function () {
@@ -205,13 +205,13 @@ describe("jasmine.Fixtures", function () {
 
     it("should not preload the same fixture twice", function () {
       jasmine.getFixtures().preload(fixtureUrl, fixtureUrl)
-      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(1)
+      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
     })
 
     it("should have shortcut global method preloadFixtures", function () {
       preloadFixtures(fixtureUrl, anotherFixtureUrl)
       jasmine.getFixtures().read(fixtureUrl, anotherFixtureUrl)
-      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(2)
+      expect(jasmine.Fixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
     })
   })
 
@@ -402,11 +402,6 @@ describe("jasmine.Fixtures using real AJAX call", function () {
   describe("when fixture contains an <script src='to/your/source'> tag", function () {
     var fixtureUrl = "fixture_with_javascript.html"
 
-    it("should load javascript in fixture", function () {
-      var fixtureContent = jasmine.getFixtures().read(fixtureUrl)
-      expect(fixtureContent).toContain('$(function (){ $(\'#anchor_01\').click(function(){ $(this).addClass(\'foo\'); }) });')
-    })
- 
     it("should load content of fixture file and javascript and bind events", function () {
       jasmine.getFixtures().load(fixtureUrl)
       $('#anchor_01').click()
@@ -421,39 +416,9 @@ describe("jasmine.Fixtures using real AJAX call", function () {
       expect($("#anchor_01")).toHaveClass('bar')
     })
   })
-
 })
 
-
-describe("jQuery matchers", function () {
-  describe("when jQuery matcher hides original Jasmine matcher", function () {
-    describe("and tested item is jQuery object", function () {
-      it("should invoke jQuery version of matcher", function () {
-        expect($('<div />')).toBe('div')
-      })
-    })
-
-    describe("and tested item is not jQuery object", function () {
-      it("should invoke original version of matcher", function () {
-        expect(true).toBe(true)
-      })
-    })
-
-    describe("and tested item is a dom object", function () {
-      it("should invoke jquery version of matcher", function () {
-        expect($('<div />').get(0)).toBe('div')
-      })
-    })
-  })
-
-  describe("when jQuery matcher does not hide any original Jasmine matcher", function () {
-    describe("and tested item in not jQuery object", function () {
-      it("should pass negated", function () {
-        expect({}).not.toHaveClass("some-class")
-      })
-    })
-  })
-
+describe("jQuery matcher", function () {
   describe("when invoked multiple times on the same fixture", function () {
     it("should not reset fixture after first call", function () {
       setFixtures(sandbox())
@@ -591,7 +556,7 @@ describe("jQuery matchers", function () {
       $("#sandbox").css("height", "auto");
       $("#sandbox").css("margin-left", "auto");
       $("#sandbox").css("display", "none");
-      expect($("#sandbox")).toHaveCss({height: 'auto', 'margin-left': "0px", display: "none"});
+      expect($("#sandbox")).toHaveCss({height: 'auto', 'margin-left': "auto", display: "none"});
     })
   })
 
@@ -942,6 +907,7 @@ describe("jQuery matchers", function () {
       expect($three.length).toBe(3)
       expect($three).toHaveLength(3)
     })
+
     it("should pass negated on an object with more than zero items", function () {
       var $three = $('<div>').add('<span>').add("<pre>")
       expect($three.length).toBe(3)
@@ -961,35 +927,35 @@ describe("jQuery matchers", function () {
     })
   })
 
-  describe("toBe", function () {
+  describe("toEqual", function () {
     beforeEach(function () {
       setFixtures(sandbox())
     })
 
     it("should pass if object matches selector", function () {
-      expect($('#sandbox')).toBe('#sandbox')
-      expect($('#sandbox').get(0)).toBe('#sandbox')
+      expect($('#sandbox')).toEqual('#sandbox')
+      expect($('#sandbox').get(0)).toEqual('#sandbox')
     })
 
     it("should pass negated if object does not match selector", function () {
-      expect($('#sandbox')).not.toBe('#wrong-id')
-      expect($('#sandbox').get(0)).not.toBe('#wrong-id')
+      expect($('#sandbox')).not.toEqual('#wrong-id')
+      expect($('#sandbox').get(0)).not.toEqual('#wrong-id')
     })
   })
 
-  describe("toContain", function () {
+  describe("toContainElement", function () {
     beforeEach(function () {
       setFixtures(sandbox().html('<span />'))
     })
 
     it("should pass if object contains selector", function () {
-      expect($('#sandbox')).toContain('span')
-      expect($('#sandbox').get(0)).toContain('span')
+      expect($('#sandbox')).toContainElement('span')
+      expect($('#sandbox').get(0)).toContainElement('span')
     })
 
     it("should pass negated if object does not contain selector", function () {
-      expect($('#sandbox')).not.toContain('div')
-      expect($('#sandbox').get(0)).not.toContain('div')
+      expect($('#sandbox')).not.toContainElement('div')
+      expect($('#sandbox').get(0)).not.toContainElement('div')
     })
   })
 
@@ -1444,7 +1410,7 @@ describe("jasmine.StyleFixtures", function () {
 
   beforeEach(function () {
     jasmine.getStyleFixtures().clearCache()
-    spyOn(jasmine.StyleFixtures.prototype, 'loadFixtureIntoCache_').andCallFake(function (relativeUrl){
+    spyOn(jasmine.StyleFixtures.prototype, 'loadFixtureIntoCache_').and.callFake(function (relativeUrl){
       this.fixturesCache_[relativeUrl] = ajaxData
     })
   })
@@ -1530,7 +1496,7 @@ describe("jasmine.StyleFixtures", function () {
       it("should go from cache", function () {
         jasmine.getStyleFixtures().preload(fixtureUrl, anotherFixtureUrl)
         jasmine.getStyleFixtures().load(fixtureUrl, anotherFixtureUrl)
-        expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(2)
+        expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
       })
 
       it("should return correct CSSs", function () {
@@ -1542,12 +1508,12 @@ describe("jasmine.StyleFixtures", function () {
 
     it("should not preload the same fixture twice", function () {
       jasmine.getStyleFixtures().preload(fixtureUrl, fixtureUrl)
-      expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(1)
+      expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
     })
 
     it("should have shortcut global method preloadStyleFixtures", function () {
       preloadStyleFixtures(fixtureUrl, anotherFixtureUrl)
-      expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.callCount).toEqual(2)
+      expect(jasmine.StyleFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
     })
   })
 
@@ -1663,7 +1629,7 @@ describe("jasmine.JSONFixtures", function () {
 
   beforeEach(function () {
     jasmine.getJSONFixtures().clearCache()
-    spyOn(jasmine.JSONFixtures.prototype, 'loadFixtureIntoCache_').andCallFake(function (relativeUrl){
+    spyOn(jasmine.JSONFixtures.prototype, 'loadFixtureIntoCache_').and.callFake(function (relativeUrl){
       fakeData = {}
       // we put the data directly here, instead of using the variables to simulate rereading the file
       fakeData[fixtureUrl] = {a:1, b:2, arr: [1,2,'stuff'], hsh: { blurp: 8, blop: 'blip' }}
@@ -1710,14 +1676,14 @@ describe("jasmine.JSONFixtures", function () {
       expect(getJSONFixture(fixtureUrl)).toEqual(ajaxData)
       expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_).toHaveBeenCalled()
       expect(getJSONFixture(anotherFixtureUrl)).toEqual(moreAjaxData)
-      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.length).toEqual(2)
+      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(2)
     })
 
     it("retrieves from cache on subsequent requests for the same fixture", function () {
       expect(getJSONFixture(fixtureUrl)).toEqual(ajaxData)
-      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.length).toEqual(1)
+      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
       expect(getJSONFixture(fixtureUrl)).toEqual(ajaxData)
-      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.length).toEqual(1)
+      expect(jasmine.JSONFixtures.prototype.loadFixtureIntoCache_.calls.count()).toEqual(1)
     })
   })